Add functionality to assign person to a veileder in the current enhet
Get all veiledere from syfoveileder, with new store-files.
Add a button that gives you a dropdown with all veiledere (as radio buttons) on the current enhet, except yourself.
A veileder can then select personer and a veileder to assign.
The dropdown has an input field, when the veileder writes in it, it will filter the list of veiledere based on fornavn, etternavn, and ident.
Since we now get veiledere from syfoveileder, remove veiledere from personRegisterState.
Update tests to reflect the changes done to different components.